♻️ refactor(admin-console): split frontend API modules
