+++ apps/admin-console/src/lib/api/index.ts
@@ -1,0 +1,11 @@
+export { agentsApi } from "./agents";
+export { auditApi } from "./audit";
+export { capabilitiesApi } from "./capabilities";
+export { configApi } from "./config";
+export { configResourceApi } from "./config-resource";
+export { ADMIN_TOKEN_STORAGE_KEY, BACKEND_URL, ConfigApiError, agentPreviewRunUrl } from "./http";
+export { mcpApi } from "./mcp";
+export { providersApi } from "./providers";
+export { systemApi } from "./system";
+export { toolsApi } from "./tools";
+export * from "./types";
